transcribe/pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = ""
authors = [
    {name = "Andre van Zuydam",email = "andrevanzuydam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch (>=2.5.1,<3.0.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "yt-dlp (>=2025.1.26,<2026.0.0)",
    "pyannote-audio (>=3.3.2,<4.0.0)",
    "pydub (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
